refactor packages and file naming semantics
