+++ src/app.module.ts
@@ -4,7 +4,7 @@ import { ScheduleModule } from '@nestjs/schedule'
 
 import { ApiModule } from '@src/module.api'
 import { DeFiDModule } from '@src/module.defid'
-import configuration from '@src/configuration'
+import configuration from '@src/app.configuration'
 
 @Module({
   imports: [
